Use SDL gamepad abstractions.
These configure common gamepads' button numbers to mappings that match
something on an xbox 360 type controller and make it so buttons labelled
"A" on different controllers will all give the same button id.  If a
controller is not in the database, however, no fallback is provided at
this time and the controller will simply be ignored.